Implement sending messages through ChatModel
Also required refactoring ChatModel to store its own information
on messages rather than relying on Tp::Message, which is a good
idea anyway.

This is currently a bit hacked in (but isn't everything?). It might
make sense later to have a class for managing conversations separate
from the model itself.