Add AccountSelector for new conversations
This whole concept (selecting the local account for a message) might
be deprecated once contacts are available - since the ideal approach
is to select which of a contact's accounts you're going to message.
Regardless, this is necessary for now to have a path to the telepathy
account, from which we can start a new channel.

AccountSelector is just a button (UI very much temporary) that opens
a SelectionDialog, fed by AccountsModel. Due to a quirk in QML, the
AccountsModel has to have a count property for SelectionDialog to
function.