Channel requests & starting conversations
First work on channel requests for starting new conversations.
This includes the UI (a new state of ConversationPage),
significant backend work to handle requests, etc. AccountsModel
and AccountSelector are currently central to the process, used
to get the Tp::Account for creating a channel; once contacts
integration is done, that concept will change dramatically.

A followup refactor to the backend cleans up this implementation
and makes ChatModel not manage telepathy state anymore.