Refactor to add ConversationChannel
ConversationChannel is a single class managing a telepathy channel
throughout its lifetime, including from pending channel requests.
It also takes over management of existing channels from ChatModel,
where it really did not belong.

This is a big refactor, but it's a much better structure for the
immediate future. There are still outstanding problems:
 - Error management isn't implemented
 - The lifetime of the ConversationChannel object is undefined
   (and thus leaky)
 - The relationship with the model is unclean in places