Buffer messages in ConversationChannel during creation
This makes it possible to send messages for a channel before it has
been fully created; those are shown in the ChatModel immediately, but
not actually sent until the channel is ready for them.

This involves a bit of ugly hacking to get them displayed in ChatModel.
I expect that to be removed at some point, when ChatModel is
separated a bit more from ConversationChannel, or when support for
message states is added.