Fix MessagesView scrolling
The attempt at attaching items to the bottom of the screen caused
some bad visual bugs (including all of the text disappearing
entirely), probably due to contentHeight/contentY being unreliable.
Disabled that for now, since it's better to have a view that works.

Also fixed the scroll-to-bottom logic to always work correctly by
using positionViewAtEnd, and to not break with the VKB open (which
was caused by contentHeight changing oddly).

This should be reliable for now; I'll revisit the bottom attaching
behavior later.