Fit text boxes to text and add highlighting
Experimental idea; rather than having the text boxes at a fixed
width (formerly 80%) and adjustable height with a minimum (60px),
fit the boxes exactly to the dimensions of the text, with wrapping.

Also added a raised style on the message text, to make it slightly
more readable when on the blue.