Implement a hackish solution for MessagesView order
Commhistory's events models are in the opposite order as what we
want to display, and QML views offer no possibility of handling that.

Short of changing the model, this is the best hack I can come up with.
Rotate the view, then rotate the delegate again, and apply some really
ugly anchoring to get ScrollDecorator in the right place.

I'm sorry.