Keep the VKB open when the Send button is clicked
This behavior requires a bit of a trick. TextArea will give up focus
and close the VKB after any outside click; that can be disabled by
wrapping the TextArea in a FocusScope, which makes giving focus to the
parent a no-op.

Then, an InverseMouseArea can be used to actually give up focus for
clicks in the desired area (in this case, anywhere outside of the
ChatTextInput).