Move C++ logic to a QML plugin
To make UI easily replaceable, this logic is better handled in
libraries. This is all moved to nemo-qml-plugin-messages now.
WindowManager cannot be moved, as it's responsible for
creating the QML view, which we can't sanely do within QML.

The ContextProvider was also removed from WindowManager, to be
implemented separately in the plugin.